feat: 회원가입 API 구현

+++ Eno/umc9th/src/main/java/com/example/umc9th/domain/member/dto/res/MyPageDto.java
@@ -14,5 +14,5 @@ public class MyPageDto {
     private Gender gender;
     private LocalDate birthDate;
     private String address;
-    private Long totalPoint;
+    private int totalPoint;
 }
